requirements.txt
streamlit
streamlit-extras
streamlit-authenticator
fastapi==0.109.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pymongo==4.6.1
python-dotenv==1.0.0
pyyaml
uvicorn==0.27.0
pydantic==2.5.3 